# PERF: Use accelerated protocol