Running this small benchmark:

```python
import pymapd
con = pymapd.connect(user="mapd", password="HyperInteractive", host="localhost", dbname="mapd")
list(con.execute("select * from flights_2008_10k"))
```

Master:

```
%timeit list(con.execute("select * from flights_2008_10k"))
3.13 s ± 23.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

HEAD

```
%timeit list(con.execute("select * from flights_2008_10k"))
1.41 s ± 11 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

So ~2x speedup for 10,000 rows.

Thanks for the tip @andrewseidl