@TomAugspurger 2017-08-21
Here's the results for `'select flight_year, flight_month, flight_dayofmonth, depdelay, uniquecarrier from flights_2008_7M'`

Master:

```
In [7]: %timeit list(con.execute(query))
4min 34s ± 665 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

HEAD

```
In [3]: %timeit list(con.execute(query))
2min 21s ± 4.91 s per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

so again, about a 2x speedup.